+++ lib/command/no_command.rb
@@ -10,6 +10,7 @@ module Command
     DESC
     HIDE = true
     WITH_INFO_HEADER = false
+    VALIDATIONS = [].freeze
 
     def call
       if config.options[:version]