+++ lib/command/setup_app.rb
@@ -15,6 +15,7 @@ module Command
       - Automatically binds the app to the secrets policy, as long as both the identity and the policy exist
       - Use `--skip-secret-access-binding` to prevent the automatic bind
     DESC
+    VALIDATIONS = %w[config templates].freeze
 
     def call # rubocop:disable Metrics/MethodLength
       templates = config[:setup_app_templates]
@@ -32,9 +33,9 @@ module Command
 
       progress.puts
 
-      if cp.fetch_identity(app_identity).nil? || cp.fetch_policy(app_secrets_policy).nil?
-        raise "Can't bind identity to policy: identity '#{app_identity}' or " \
-              "policy '#{app_secrets_policy}' doesn't exist. " \
+      if cp.fetch_identity(config.identity).nil? || cp.fetch_policy(config.secrets_policy).nil?
+        raise "Can't bind identity to policy: identity '#{config.identity}' or " \
+              "policy '#{config.secrets_policy}' doesn't exist. " \
               "Please create them or use `--skip-secret-access-binding` to ignore this message." \
               "You can also set a custom secrets name with `secrets_name` " \
               "and a custom secrets policy name with `secrets_policy_name` " \
@@ -42,7 +43,7 @@ module Command
       end
 
       step("Binding identity to policy") do
-        cp.bind_identity_to_policy(app_identity_link, app_secrets_policy)
+        cp.bind_identity_to_policy(config.identity_link, config.secrets_policy)
       end
     end
   end