+++ lib/core/controlplane.rb
@@ -3,6 +3,8 @@
 class Controlplane # rubocop:disable Metrics/ClassLength
   attr_reader :config, :api, :gvc, :org
 
+  NO_IMAGE_AVAILABLE = "NO_IMAGE_AVAILABLE"
+
   def initialize(config)
     @config = config
     @api = ControlplaneApi.new
@@ -36,6 +38,51 @@ class Controlplane # rubocop:disable Metrics/ClassLength
   end
 
   # image
+
+  def latest_image(a_gvc = gvc, a_org = org, refresh: false)
+    @latest_image ||= {}
+    @latest_image[a_gvc] = nil if refresh
+    @latest_image[a_gvc] ||=
+      begin
+        items = query_images(a_gvc, a_org)["items"]
+        latest_image_from(items, app_name: a_gvc)
+      end
+  end
+
+  def latest_image_next(a_gvc = gvc, a_org = org, commit: nil)
+    commit ||= config.options[:commit]
+
+    @latest_image_next ||= {}
+    @latest_image_next[a_gvc] ||= begin
+      latest_image_name = latest_image(a_gvc, a_org)
+      image = latest_image_name.split(":").first
+      image += ":#{extract_image_number(latest_image_name) + 1}"
+      image += "_#{commit}" if commit
+      image
+    end
+  end
+
+  def latest_image_from(items, app_name: gvc, name_only: true)
+    matching_items = items.select { |item| item["name"].start_with?("#{app_name}:") }
+
+    # Or special string to indicate no image available
+    if matching_items.empty?
+      name_only ? "#{app_name}:#{NO_IMAGE_AVAILABLE}" : nil
+    else
+      latest_item = matching_items.max_by { |item| extract_image_number(item["name"]) }
+      name_only ? latest_item["name"] : latest_item
+    end
+  end
+
+  def extract_image_number(image_name)
+    return 0 if image_name.end_with?(NO_IMAGE_AVAILABLE)
+
+    image_name.match(/:(\d+)/)&.captures&.first.to_i
+  end
+
+  def extract_image_commit(image_name)
+    image_name.match(/_(\h+)$/)&.captures&.first
+  end
 
   def query_images(a_gvc = gvc, a_org = org, partial_gvc_match: nil)
     partial_gvc_match = config.should_app_start_with?(a_gvc) if partial_gvc_match.nil?