+++ spec/support/dummy_app_setup.rb
@@ -4,16 +4,7 @@ module DummyAppSetup
   module_function
 
   def setup
-    config_dir = "#{CommandHelpers.spec_directory}/dummy/.controlplane"
-
-    config_file = File.read("#{config_dir}/controlplane.yml")
-    config_file = config_file.gsub("{GLOBAL_IDENTIFIER}", CommandHelpers.dummy_test_app_global_identifier)
-
-    @tmp_config_file = Tempfile.create(["controlplane-tmp-", ".yml"], config_dir)
-    @tmp_config_file.write(config_file)
-    @tmp_config_file.rewind
-
-    ENV["CONFIG_FILE_PATH"] = @tmp_config_file.path
+    CommandHelpers.configure_config_file
 
     puts "\nUsing org '#{CommandHelpers.dummy_test_org}' for tests with dummy app\n\n"
   end
@@ -27,6 +18,6 @@ module DummyAppSetup
       end
     end
 
-    File.delete(@tmp_config_file.path) if @tmp_config_file
+    CommandHelpers.delete_config_file
   end
 end